// src/app/Models/ProductoDetalle.ts
export interface ProductoDetalle {
    cia: string;
    anio: string;
    mes: string;
    producto: string;
    secuencia: any;
    cantidad: any;
    canal_codigo: string;
    canal_nombre: string;
    cliente_codigo: string;
    cliente_nombre: string;
    ensamblador: string;
    ensamblador_nombre: string;
    estado: string;
    area_codigo: string;
    area_trabajo: string;
    flag_sin_insumos: number;
    obs: string;
    advertencia: string;
}